netconsd: Replace probing hashtable with rculfhash from liburcu
Netconsd turns eight soon! It deserves some love.

This replaces the ugly per-thread probing hashtables with one global
instance of the concurrent hashtable from liburcu. It also uses the
linked list implementation from liburcu for the timerlists, deleting
a bunch of code in the process.

In addition to improving performance, the rcu hashtable allows the
garbage collection to actually be asynchronous in its own thread.

Messages are still consistent hashed to workers in the same way.

Signed-off-by: Calvin Owens <calvin@wbinvd.org>